Move `WooTabNavigationController` to its own file and put large titles behind the feature flag.
